Add files via upload
These are the registrations for the employee, organization, and trainer, also included is the logout page when a user is done with what they doing they need to log out. The db connect is just to establish a connection to the database. It doesnt have to be separate but i just like to separate it. We will have to change some of the names once we create our relations and attributes

am working on the referential integrity constraints and the indexing. need to read first.

Guys you are most welcome to make changes where ever necessary