login.php
I think when our users login, we should have a query that goes to the db to check if that user is already in the db and if they are then redirect them to their respective pages. We will need to improve this because upon checking the radio button then it should know where to direct to. we will talk next meeting and find means of improving it. We need three more pages. Employee with all the necessary detail and same for both trainee and organization